mentor_agent/requirements.txt
pydantic>=2.0.0
requests>=2.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
